refactor(authentication system): consolidate authentication modals into a single component
